Fixed some messages not resulting in status :done
Signed-off-by: Александар Симић <asimic@gmail.com>

Fix for :status :done

The original patch was submitted by
Dirk Geurs <dirklectisch@mac.com> on 30/Mar/15 10:26 AM

His description of the problem:

According to the documentation:

"Once a handler has completely processed a message, a response
containing a `:status` of `:done` must be sent."

I am assuming this includes messages that have resulted in an
error. Some messages to an nREPL server never respond with such a
status possibly leaving some clients waiting for follow up messages
forever.

So far I have found two instances of this problem.

1. The interruptible-eval middleware may result in a :no-code error
that does not include the :done status.

2. The session middleware may result in a :unknown-session error that
does not include the :done status.

I'm including a patch for the above two which are just one word
fixes. There might be other situations in which the problem arises.