[Fix #16] ThreadPoolExecutor constructor is now created with corePoolSize = 1 instead of 0 (#36)
I think we have a rather logical, expected, and flexible behavior now:

1. When the session is started and the first connection to it made, the session does get one permanent thread for evaluation. It is mapped to this same permanent thread for its entire existence.
2. When another connection gets attached to the same session, its evaluation do happen in the default permanent thread, unless that thread is busy, in which case new thread is being created.
3. Executor thread provides threads per individual evaluations, but uses the pool, in which there is always at least default thread already available, which is used preferably.
4. As it is now, the only case where the default session thread is not used for evaluation is when there is already evaluation running in the same session. I think this is a) rare b) expected, since the alternative is that all connections to the same session block and wait until the first finish with its evaluation.

The change is trivial. Maybe in the future the whole function could be replaced by one of the Java's thread pool executor factory methods that does the same. It seems that one of many factory methods does this, but I opted to keeping the change minimal at first until we are sure that this is really the solution in more exotic cases...

See https://github.com/nrepl/nREPL/pull/36 for more details.