Remove some legacy Java connection API
This was never documented and I couldn't find why was it created in
the first place. Seems it added so it'd be easier to connect to
nREPL from a native Java app.

I don't think that anyone actually uses it given its undocumented
state, so we're better off without it.