Revert "Extract version from project metadata"
This reverts commit 76c9a0b163e62b0e59dd523fe9903321c3174398.

I had borrowed this approach from other projects, but I realized
that while it works ok for libraries, it's not a great option for
nREPL when it's used with `clj`. When I was trying the dev version
of nREPL I noticed that now it was picking up for some reason the
project.clj of the projects I was starting nREPL in. I'm not quite
sure why this happens, but I guess it's best to revert this for now.