Show `:host` and `:bind` option keywords usage
Documentation currently doesn't mention how to make nREPL server
listen on a particular address. And the only documentation talking
about the host address is shown for the command line client, where the
`--host` option is mentioned. This might mistake people into thinking
that the corresponding option for the server starting function also
uses an option with the same keyword name. Which is not the case, as
the option keyword is actually named `:bind`.

Add a couple of sentences documenting the option keywords both for
the client and the server side functions and code snippets, as
suggested by @bbatsov.