Fix `run-repl` using a different transport
When testing with fastlane transports, the REPL was being
started using bencode and was giving an "The transport's socket
appears to have lost its connection to the nREPL server" error.

We add a `transport` key to `run-repl` args so it can be
connect to the server using the right transport.