Allow suppressing ack message when using nrepl.cmdline
This commit adds a `--verbose` command line option, and only displays the ack
message if this flag is passed. It also prints the message to `*out*` rather
than `*err*`, which didn't seem to make sense since it's not an erroneous
condition. Finally, the previous message also included the result of calling
`send-ack`, which always returns `nil`.

Fixes https://github.com/nrepl/nrepl/issues/167