[Fix #113] Change how we set classloaders to fix hotloading in Lein (#190)
As far as I can tell, the cause is that the combination of using Leiningen in/out of projects, as stated in the original issue, is causing a variation in when `Compiler/LOADER` is first bound. If this happens early enough, the execution thread's `ContextualClassLoader` might not have a suitable common ancestor with `Compiler/LOADER` for Pomegranate to modify. 

This change skips the LOADER value at `eval` time, using instead, the `ContextualClassLoader` instead. This is quite likely the `DynamicClassloader` for the thread, set by the `session` middleware. This guarantees the common ancestor property.

I added a check to see the two classloaders have a common ancestor in `eval`.

I'm a bit worried about this breaking something in a very complex Java environment with exotic class loading behaviour that depends on specific interaction with how the Clojure compiler behaves, so would be good to get at least some testing by someone who works on such a stack.