Retain file meta when file is not in classpath
By default, the :file meta is the path to the file relative to the
classpath root. However, many nREPL clients send
^{:clojure.core/eval-file ,,,} metadata in eval ops. As a result, the
:file meta of vars defined this way becomes the file path the client
sends in that metadata map.

In practice, the file path clients send often seems to be the absolute
path to the file. Trying to resolve that path against the classpath
fails, which results in the lookup op returning a :file key with an
empty string as a value.

After this commit, the code first tries to look up the value of :file on
the classpath. If that fails, it retains the :file key as is.