Fix lookup of non-special symbol special forms
"let" is a special form, but not a special symbol. That is, the
"special-symbol?" fn returns false on "let".

("let*" is the special symbol that backs "let".)

A lookup on "let" returns the :special-form key. Its value is a
boolean, which cannot be Bencoded. We therefore convert the boolean
into a string first, the same way as with the :macro key.