Fix sideloader race condition
Prior to this change, the sideloader sometimes incorrectly sent the
client an "unexpected-provide" response when the system was under high
CPU load and the number of sideloader requests was high.

The root cause was that the sideloader sent the client the lookup
request before swapping the request into the atom that contained pending
requests. As a result, the client sometimes sent the sideloader-provide
response *before* the server -- presumably blocked by contention -- had
finished swapping the request into the atom.