Store sideloader state in the session
Pending resources in the sideloader are kept in an atom, which currently is
defined in a closure which closes over the classloader and sideloader ops. When
the middleware stack is rebuilt (e.g. via an `add-middleware` op), this atom
gets re-initialized, and the state is lost, causing the side loader to respond
to previously requested resources with `:unexpected-provide`.

The only state that does persist is the session. We already use the session to
store the classloader itself, it makes sense to also keep the `pending` atom in
the session, so that the pending state that we see always corresponds with the
current session classloader.

This fixes an issue which was pointed out in the PR that adds dynamic middleware
loading: https://github.com/nrepl/nrepl/pull/185#issuecomment-624317912