Honor the session classloader when dynamically loading print-fn
The client can request that the print middleware uses a specific function for
printing. If this var is not yet defined it will be required+resolved. When
using the sideloader this might have to be loaded via the sideloader, so we need
to make sure we use the session-specific classloader.