Introduce a session-aware version of requiring-resolve
This introduces `misc/session-resolve`, which is like `requiring-resolve`, but
takes and extra session argument, and will use the session classloader to the
requiring, if the initial resolve fails. This makes it more conservative than
with-session-classloader+requiring-resolve, since the classloader will only be
created and set when requiring needs to happen.