Sideloader: filter by prefixes, cache results
Cache previous sideloader lookups, so a resource/class is only fetched once. We
do this by keeping results around in the `pending` atom (renamed to `loaded`),
and instead of storing the bytes we store the Class instance or URI to the
resource file, to prevent memory bloat.

Allow passing an optional list of `prefixes` to the sideloader, which will be
used to determine whether a certain resource/class lookup should be attempted
via the sideloader.

Both Clojure and ClojureScript cause a lot of class/resource lookup missesi. E.g.
Clojure will first look for a `__init.class` resource and a `__init` class,
and only when those are not found does it look for the clj file. If we run all
these lookups through the sideloader this causes an unacceptable hit to
performance, so clients are encouraged to tell us beforehand what they can
provide. Note that in extremis a client can simply send the full list of files
on its own "classpath", but it can also send a list of discrete namespace
prefixes that it knows about.