safe-write-bencode: unify OutputStreams and channels via Writable
The (.write ...) has to handle both OutputStreams (junixsocket) and
BufferedOutputChannels (JDK16), and so the ^OutputStream hint was
causing a crash when the latter were in use:

  ERROR: Unhandled REPL handler exception processing message {:id init, :op clone}
  java.lang.ClassCastException: class nrepl.socket.BufferedOutputChannel cannot be cast to class java.io.OutputStream (nrepl.socket.BufferedOutputChannel is in unnamed module of loader clojure.lang.DynamicClassLoader @3af356f; java.io.OutputStream is in module java.base of loader 'bootstrap')
          at nrepl.transport$safe_write_bencode.invokeStatic(transport.clj:112)
          at nrepl.transport$safe_write_bencode.invoke(transport.clj:104)
          at nrepl.transport$bencode$fn__673.invoke(transport.clj:131)
          at nrepl.transport.FnTransport.send(transport.clj:35)

Fix the problem by changing Writable to a protocol, extending
OutputStream to support it, and using it (socket/write) instead of the
OutputStream (.write ...).

Thanks to Shogo Ohta for reporting the problem.

cf. https://github.com/nrepl/nrepl/issues/245