Extend timeouts in cmdline tests to accommodate slower hosts
The tests were failing on the Debian arm64 buildd's, and it turns out
that's because those machines take longer to start the subprocesses
that the tests need.  To allow the tests to succeed, increase the wait
times in cmdline-test and cmdline-tty-test, and rework the tty-server
test to be able to detect when the server is available instead of just
sleeping unconditionally.  That should also allow the test to run
faster on faster machines.

Thanks to Louis-Philippe Véronneau for reporting the problem.