Fix bugged TLS accept loop logic (#307)
TLS server accept loop should continue in the presence of non-SSL exceptions such as SocketException. The TLS accept method will now wrap any non SSL exceptions in SSLException. The server accept loop will close the server socket upon encountering an unexpected exception. This was not the case earlier, and this lead to a lingering, open server socket where the OS would continue to accept new connections. It seems that .connect, which includes a connect-timeout-ms parameter, on a TLS socket does *not* include the handshake. Thus, in the tests, if the server first encountered a SocketException and this lead to a lingering socket, the next TLS connect would succeed, but the handshake would never finish (and there was no timeout specified). Thus, the test became stuck given a non SSL exception, and caused the CI system to time out after 10 minutes.