Add and populate :host and :unix-socket keys on nrepl.server.Server
Also updates existing code to expect the socket path on the :unix-socket
field, rather than :socket.

This fixes an issue that occurs when using the `--interactive` option in
concert with `--socket PATH`. Since the Server object has no :unix-socket
key, nrepl.cmdline/interactive-repl won't recognize that it should
connect to the socket and will exit with an error message.

This also ensures that consumers of the record can obtain the passed-in
host value.